Move get-content logic into pipe
