+++ src/app/shared/get-content.pipe.ts
@@ -1,19 +1,62 @@
 import { Pipe, PipeTransform } from '@angular/core';
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 
-import { ContentsXmlService } from '@core/contents-xml.service';
+import { Observable, of } from 'rxjs';
+import { catchError, map } from 'rxjs/operators';
+
+import { ProductContentPipe } from '@shared/product-content.pipe';
 
 @Pipe({
   name: 'sharedGetContent'
 })
 export class GetContentPipe implements PipeTransform {
-  constructor(public contentService: ContentsXmlService) {}
+  constructor(public httpClient: HttpClient) {}
 
-  transform(product, content): any {
+
+  /**
+   * Download product content
+   *
+   * @param product
+   *     Product with contents property
+   * @param content
+   *     Name of the content to download
+   */
+  getContent (product: any, content: string): Observable<any> {
+    const pcPipe = new ProductContentPipe();
+    const _content = pcPipe.transform(product, content);
+
+    if (!_content) {
+      of(null);
+    }
+
+    try {
+      return this.httpClient
+        .get(_content.url)
+        .pipe(map(response => response))
+        .pipe(catchError(this.handleError()));
+    } catch (e) {
+      return of(null);
+    }
+  }
+
+  transform (product, content): any {
     if (!product || !content) {
       return null;
     }
 
-    return this.contentService.getContent(product, content);
+    return this.getContent(product, content);
   }
 
+
+  /**
+   * Error handler for http requests
+   *
+   * @returns
+   *    returns error
+   */
+  private handleError() {
+    return (error: HttpErrorResponse): Observable<any> => {
+      return of(null);
+    };
+  }
 }
